temp_deploy/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
requests==2.31.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.100.0
aiohttp==3.8.5
gunicorn==21.2.0
Werkzeug==2.3.7
